Added basis for the lobby class and its server-side functionality
Within this commit, I added a lobby class that holds 3 variables: the id of the lobby, the name of the lobby, and an ArrayList of the users' ids that currently partake in said lobby. Furthermore, within this lobby class, one may add new users via the addNewUsers function, which updates the users arraylist to include the user's id.