Fixed issue with the server-side tests for lobby
When attempting to incorporate the function getLobby into a test, it returned an error that it couldn't parse through the context given. To fix this, I made this branch to check if the issue occurred only on my computer. Since it did not, I went onto one of the lab computers and fixed the error. Also, to ensure a greater amount of coverage, I added server side testing for deleting a Lobby.